requirements.txt
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-WTF==0.14.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
PyJWT==1.7.1
python-dotenv==0.10.3
pytz==2019.3
requests==2.22.0
six==1.13.0
twilio==6.35.2
urllib3==1.25.7
Werkzeug==0.16.0
WTForms==2.2.1
